+++ backend/coffeeing/src/main/java/com/ssafy/coffeeing/modules/product/domain/Coffee.java
@@ -48,7 +48,7 @@ public class Coffee extends BaseEntity {
 
     @Column
     @Builder.Default
-    private Double totalScore = 0.0;
+    private Integer totalScore = 0;
 
     @Column
     @Builder.Default
@@ -63,7 +63,7 @@ public class Coffee extends BaseEntity {
         this.totalReviewer++;
     }
 
-    public void deleteReview(Double score) {
+    public void deleteReview(Integer score) {
         this.totalScore -= score;
         this.totalReviewer--;
     }
